backend/requirements.txt
Flask==3.1.0
pymongo==3.11.3
bcrypt==3.2.0  # Update bcrypt to the latest version
flask-cors==5.0.0  # Update to the latest version of flask-cors
PyJWT==2.1.0
